// src/administrador/infraestructura/dto/oferta/VerDetalllesOfertaLaboralAdministrador.api.dto.ts
export interface VerDetallesOfertaLaboralAdministradorAPIDTO {
  uuid: string
  titulo: string
  fechaPublicacion: string
  fechaModificacion: string
  cargo: string
  sueldo: number
  descripcion: string
  duracionEstimadaValor: number
  duracionEstimadaEscala: string
  turnoTrabajo: string
  numeroVacantes: number
  uuidEmpresa: string
  empresaNombre: string
  direccionEmpresa: string
}